backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.62.1
django-bootstrap4==24.3
django-clearcache==1.2.1
django-crispy-forms==2.1
djangorestframework==3.15.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-kms==2.21.4
googleapis-common-protos==1.63.0
googlemaps==4.10.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
oauthlib==3.2.2
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
python3-saml==1.16.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
xmlsec==1.3.14
